Adding & removing switches!
Added "monday" to the names of all Monday-specific things, moved it to
the Monday activity instead of Day because I don't think they can be
generalised easily. Rewrote removeSwitch() method, implemented removing
switches and adding switches as well as Remove All functionality.
Updated to do list because there are bugs in all of it :(